+++ src/model/patching_channel.py
@@ -1,17 +1,17 @@
 # coding=utf-8
-"""Channels for patching"""
+"""Channels for patching_mode"""
 from PySide6 import QtCore
 
 from ofl.fixture import UsedFixture, Mode
 
 
 class PatchingChannel(QtCore.QObject):
-    """Channels for patching to reduce redundancy over scenes"""
+    """Channels for patching_mode to reduce redundancy over scenes"""
     updated_fixture: QtCore.Signal = QtCore.Signal()
     updated_color: QtCore.Signal = QtCore.Signal(str)
 
     def __init__(self, channel_address: int, color: str):
-        """Constructs a patching channel."""
+        """Constructs a patching_mode channel."""
         super().__init__()
         if 0 > channel_address or channel_address > 511:
             raise ValueError(f"Tried to create a channel with address {channel_address}")
