+++ src/model/patching_universe.py
@@ -21,7 +21,7 @@ class PatchingUniverse:
 
     @property
     def patching(self) -> list[PatchingChannel]:
-        """List of all 512 patching channels belonging to the Universe"""
+        """List of all 512 patching_mode channels belonging to the Universe"""
         return self._patching
 
     @property
