+++ src/view/patching_mode/patch_item.py
